Tip Calculator
This is a simple project called "Tip Calculator" made using Python language. The main purpose is just to create a program where I can calculate the total price that each person is going to pay when they decide to split the bill. It's really handy when friends who went out for dinner manage to split the bill equally. Download the code and make the test by running and customizing.